combine duplicate availability formatting function
