+++ src/EntryGridCtrl.js
@@ -5,40 +5,11 @@ import lang, { generalLangProxy } from "./resources";
 import { useCollection } from "./useCollection";
 import { useCurrentHour, useCurrentMonth } from "./helpers/useCurrentTime";
 import {
-    getTimeRanges,
-    getMonthRanges,
     willLeaveThisMonth,
     willLeaveThisHour,
+    humanReadableAvailabilityTimes,
 } from "./helpers/availabilityTimeHelpers";
 import { mappingIdToFilename } from "./collectionDataAcnhApiMapping";
-
-// duplicate from EntryCtrl.js (but changed)
-function formatAvailabilityTimes(months, hours) {
-    const ms = getMonthRanges(months)
-        .map((rng) =>
-            rng.allYear
-                ? lang.availability.ALL_YEAR
-                : rng.from === rng.to
-                ? lang.availabilityShort[rng.from]
-                : lang.availabilityShort[rng.from] +
-                  " " +
-                  lang.availability.TO +
-                  " " +
-                  lang.availabilityShort[rng.to]
-        )
-        .map((text) => ({ text, type: "MONTH" }));
-    const hs = getTimeRanges(hours)
-        .map((rng) =>
-            rng.allDay
-                ? lang.availability.ALL_DAY
-                : rng.from + "-" + rng.to + " " + lang.availability.CLOCK
-        )
-        .map((text) => ({ text, type: "TIME" }));
-
-    if (ms && hs) return ms.concat(hs);
-
-    return ms || hs;
-}
 
 export default function EntryGridCtrl({
     entry: {
@@ -101,22 +72,23 @@ export default function EntryGridCtrl({
                             )}
                             {location && <span>{lang.location[location]}</span>}
 
-                            {formatAvailabilityTimes(northernMonths, hours).map(
-                                (availability) => (
-                                    <span
-                                        className={
-                                            (leavingThisMonth &&
-                                                availability.type ===
-                                                    "MONTH") ||
-                                            (leavingThisHour &&
-                                                availability.type === "TIME")
-                                                ? "highlight"
-                                                : ""
-                                        }>
-                                        {availability.text}
-                                    </span>
-                                )
-                            )}
+                            {humanReadableAvailabilityTimes(
+                                northernMonths,
+                                hours,
+                                true
+                            ).map((availability) => (
+                                <span
+                                    className={
+                                        (leavingThisMonth &&
+                                            availability.type === "MONTH") ||
+                                        (leavingThisHour &&
+                                            availability.type === "TIME")
+                                            ? "highlight"
+                                            : ""
+                                    }>
+                                    {availability.text}
+                                </span>
+                            ))}
                         </>
                     )}
                 </div>
